Add some frontend pages with BrowserRouter
- NotFound: simple 404 Error page
- Robot: Set up skeleton (so we can call the param id)

Todo: We still need to query the API (which queries the backend) for
Robot. We are blocked on

1) what datatype are we serializing the information from the Robot table
  to? (We ought to make this datatype in the frontend too, i.e. in some
  typescript file)
2) What API path allows us to get information about a robot? This is an
  easier function to write because we just need to fetch info from db,
  everyone sees the same info since it's public.
3) how does the frontend interact with the API for this page? A little
  easier, just throw a request at the correct route, properly serialize
  the data by fitting it to the correct typescript type. But blocked by
  1) and 2) atm

and as a sort of separate concern, where are we storing images? And how
do we store which images are associated with each robot? Here's what I'm
thinking right now: chuck some images in some AWS bucket, serve the
images publicly (I presume no images need to be private), and what each
entry in the robots table stores is instead a list of URLs to the
images.)