remove redis (#160)
* remove redis

* continuing work-in-progress

* move some stuff around

* more wip

* misc clean uip

* Refactor get_user_id_from_session_token into get_user_from_jwt

The function get_user_from_jwt is unimplemented but I have left a
comment stating how it should be implemented. In particular, we ought to
throw an error if the JWT is not associated with a user.

Furthermore, I have replaced every instance of user_id with id.

* Fix circular import with TABLE_NAME

* replace new_token fns with appropriate jwt generation placeholders

* unify oauthuser with normal user model

* Revert "replace new_token fns with appropriate jwt generation placeholders"

We are not using JWT's. We are going to use API keys anyway, which makes
JWTs completely pointless.

This reverts commit b66d8ad0c9d177c429555b9d2cb31bf308491137.

* remove calls to get_api_key fn that isn't doing anything

* get rid of jwts for good and format stuff

* fix all mypy errors except for weird GitHub id thing

almost done with boilerplate

* fix lints for real by adding new custom exception

* Make email field optional

* some stuff

* model schema changes

* clean up

* clean up some stuff

* Just have type of auth_keys be list[str] for User

No one is going to voluntarily pass in auth_keys=None. We ought to just
make the default value the empty list rather than None, since that is
how it is passed into the constructor in the first place

* Get rid of email-password login.

We are transitioning into using OAuth exclusively.

* Remove frontend pages associated with email/password login

* Fix accidental change of GitHub login URL

* Almost-working everything except User creation in model.py

* Fix OAuth GitHub

* Re-implement parts query

* Lints

* Add /robots route by genericing /parts route

* format

* Actually add robot routers, oops.

* Delete unnecessary items_per_page in robots.py

We moved it to BaseCrud

* actually call the right function oops

* Restore image uploading functionality

* Make OAUTH Keys work

* lints and format

* Remove frontend sidebar/api cruft

We no longer use email/password sign-in on the backend; change the
frontend to reflect that.

* Fix image display in frontend

* Fix user owner display

Involves fixing batch query too.

* dont accidentally delete user id WHEN EDITING ROBOT

* Properly display listing owner information

* Adjust misleading router function names

* fix Your Robots/Your Parts queries

* remove unused Sidebar.tsx imports

* Split up GitHub router's GitHub API calls

This is solely to make it possible to mock them in unit tests.

* format

* Make unit tests work with new auth system/backend

* remove bespoke field_info stuff

It was causing bugs /shrug

* Add lints to tests (and have tests pass them too)

* fix lint

---------

Co-authored-by: Dennis Chen <dchen@mathadvance.org>